implement median aggregator function for sqlite backend
The median aggregator function provides more accurate predictions for
progress durations and overall progressions in most cases as it is less
likely to be thrown of by extreme outliers in the analyzed data.

AVG will still be used as the default aggregate function for the sqlite
backend as there is no native median function in SQLite and the
implementation is based on a portable but non-native python aggregate
extension.