improve expected durations and progress predictions
Up until now the expected duration was the sum of durations of all
recorded phases and the progress just the ratio of the current duration
in relation to that. This had some obvious drawbacks as the expected
duration would not change even if a phase took notably longer than any
empirical data would suggest and progress would continually increase
until it got stuck at 100%.

The new algorithm tries to improve on this behaviour. The expected
duration will now stay ahead of the reported duration as any any delays
in the process will be added to it. The progress will freeze in place if
a phase took longer than it emperically did, so that a delay in one
phase will not impact the overall progress of the process.

Apart from that the new `is_stuck` property has been added to signal the
client that the current phase has taken a notably longer time to finish
compared to empirical data.