add predictable order to phases
In 4f5eec92bbb7fbf664b4fe1d1ce865be10e5c26b the passed_phases and
expected_phases attributes were changed to contain not the number
of passed and expected phases but the actual phase names. In order to
make this more useful the phases are now in a predictable order,
or to be exact they should be in the order the phases appear in the
calling code. This will make it easier to display some kind of process
task list that shows progress for each individual step and not just
the overall progress.