fix progress calculation for unknown phases
When training docbrown for the first time we already return None when
asked for the current progress, as there aren’t any predictions to be
made.

In case a process has already been recorded and there are timings
 	available but a new phase has been added, that has not been
encountered before docbrown threw an exception until now. Instead we
allow the progress calculation but assume a runtime of zero seconds for
it.